apis: replace unsigned integers in CRD fields with signed integers, add validation
This commit replaces the unsigned integers in the HTTPProxy and IngressRoute CRDs
with signed integers. Unsigned integers are discouraged due to API compatibility
issues. Validation rules have been added to the fields to ensure that the value
provided is always positive.

Fixes #2123

+++ internal/dag/builder.go
@@ -806,7 +806,7 @@ func (b *Builder) computeRoutes(sw *ObjectStatusWriter, proxy *projcontour.HTTPP
 			c := &Cluster{
 				Upstream:              s,
 				LoadBalancerPolicy:    loadBalancerPolicy(route.LoadBalancerPolicy),
-				Weight:                service.Weight,
+				Weight:                uint32(service.Weight),
 				HealthCheckPolicy:     healthCheckPolicy(route.HealthCheckPolicy),
 				UpstreamValidation:    uv,
 				RequestHeadersPolicy:  reqHP,
@@ -1004,7 +1004,7 @@ func (b *Builder) processIngressRoutes(sw *ObjectStatusWriter, ir *ingressroutev
 				r.Clusters = append(r.Clusters, &Cluster{
 					Upstream:           s,
 					LoadBalancerPolicy: service.Strategy,
-					Weight:             service.Weight,
+					Weight:             uint32(service.Weight),
 					HealthCheckPolicy:  ingressrouteHealthCheckPolicy(service.HealthCheck),
 					UpstreamValidation: uv,
 					Protocol:           s.Protocol,
